Dear lord,
Where to start.

Utils is only loaded and used by the background script.
There are some hacky callbacks to get queries to work and whatnot so that we know that the database is connected and all that jazz.

NOTE TO SELF:
Don't be retarded and use similar names for variables even in totally different functions, remember, JS's scoping is fucked up and likes to capture random variables in the env due to global variables. Might as well be programming Emacs Lisp.